refactor: split utils to pdf and html libs
